+++ synth_lut.h
@@ -39,7 +39,7 @@ typedef struct synth_rec {
  * @param channel The channel to enable calibration on
  * @return        0 on success or an errno value on error
  */
-int synth_lut_enable(const bool tx, const size_t channel);
+int synth_lut_enable(const bool tx, const uint32_t channel);
 /**
  * Enable usage of synthesizer calibration tables when setting frequency
  * for all channels.
@@ -51,7 +51,7 @@ int synth_lut_enable_all();
 /**
  * Enable usage of synthesizer calibration tables when setting frequency.
  */
-void synth_lut_disable(const bool tx, const size_t channel);
+void synth_lut_disable(const bool tx, const uint32_t channel);
 
 /**
  * Disable usage of synthesizer calibration tables when setting frequency
@@ -72,13 +72,13 @@ int synth_lut_enable_all_if_calibrated();
  *
  * @return true when enabled, false otherwise.
  */
-bool synth_lut_is_enabled(const bool tx, const size_t channel);
+bool synth_lut_is_enabled(const bool tx, const uint32_t channel);
 
 /**
  * Clear synth calibration tables for one channel.
  * This will call synth_lut_disable() internally.
  */
-void synth_lut_erase(const bool tx, const size_t channel);
+void synth_lut_erase(const bool tx, const uint32_t channel);
 
 /**
  * Clear all synth calibration tables. This will call synth_lut_disable()
@@ -93,7 +93,7 @@ void synth_lut_erase_all();
  * @param channel the channel (number, i.e. 0, rather than 'A')
  * @return        true if calibration data exists, otherwise false.
  */
-bool synth_lut_is_calibrated(const bool tx, const size_t channel);
+bool synth_lut_is_calibrated(const bool tx, const uint32_t channel);
 
 /**
  * Get the calibrated synthesizer settings for a specific frequency on a
